requirements.txt
Click>=7.0,<8.0
graphqlclient>=0.2.4,<0.3.0
toml>=0.10.0,<0.11.0
jsonpath-rw>=1.4.0,<1.5.0
PyGithub>=1.40,<1.41
requests>=2.21.0,<2.22.0
hvac>=0.7.0,<0.8.0
ldap3>=2.5.2,<2.6.0
anymarkup>=0.8.0,<0.9.0
python-gitlab>=1.11.0,<1.12.0
semver>=2.8.0,<2.9.0
python-terraform>=0.10.0,<0.11.0
boto3>=1.9.0,<=1.10.0
botocore>=1.12.159,<=1.13.0
urllib3>=1.21.1,<1.25.0
slackclient>=1.3.2,<1.4.0
pypd>=1.1.0,<1.2.0
jenkins-job-builder>=2.10.1,<2.11.0
Jinja2>=2.10.1,<2.11.0
jira>=2.0.0,<2.1.0
pyOpenSSL>=19.0.0,<20.0.0
ruamel.yaml>=0.16.5,<0.17.0
terrascript>=0.6.1,<0.7.0
tabulate>=0.8.6,<0.9.0
manifest-bouncer>=0.1.0,<0.2.0
